requirements-dev.txt
pytest==8.2.0
pytest-asyncio==0.23.7
jsonschema==4.22.0
httpx==0.27.0
python-jose==3.3.0
ruff==0.4.5
black==24.4.2
mypy==1.10.0
pip-audit==2.7.2
bandit==1.7.8
pytest-cov==5.0.0
